feat: remove deno_core, migrate built-in interceptors to Node.js (Phase 3)
Built-in interceptors (add-header, auth-apikey, validate-request,
validate-response) are now plain CJS Node.js modules in
node-runtime/interceptors/. They are loaded by the same IPC server used
for user-defined interceptors and plugins, using ExternalRuntime.

opengateway-js-runtime is now a thin crate exporting only:
- PluginRuntime trait
- InterceptorPermissions
- CallOutput, JsBody, JsError types
- external module (ExternalRuntime, spawn, spawn_sync, locate_interceptor)

Removed:
- deno_core / V8 embedding (worker.rs, loader.rs, ops.rs, permissions.rs)
- All deno_* Cargo dependencies (~15 packages)
- interceptors/ TypeScript/esbuild build pipeline
- gateway-core/js/ bundled JS dist directory
- Dockerfile js-builder stage

module_resolver.rs now resolves internal:* module names to file paths
in node-runtime/interceptors/ at runtime via locate_interceptor().

Also adds ajv to node-runtime for JSON Schema validation.
