+++ Dockerfile
@@ -1,11 +1,3 @@
-FROM node:22-bookworm-slim AS js-builder
-
-WORKDIR /usr/src/opengateway/interceptors
-COPY interceptors/package.json interceptors/package-lock.json ./
-RUN npm ci
-COPY interceptors/ ./
-RUN npm run build
-
 FROM node:22-bookworm-slim AS node-runtime-builder
 
 WORKDIR /usr/src/opengateway/node-runtime
@@ -26,29 +18,23 @@ COPY Cargo.toml Cargo.lock ./
 COPY gateway-core/ gateway-core/
 COPY openapi-overlay/ openapi-overlay/
 COPY opengateway-js-runtime/ opengateway-js-runtime/
-COPY --from=js-builder /usr/src/opengateway/gateway-core/js/dist/ gateway-core/js/dist/
+COPY opengateway-sandbox/ opengateway-sandbox/
 RUN cargo chef prepare --recipe-path recipe.json
 
 FROM chef AS builder
 
 COPY --from=planner /usr/src/opengateway/recipe.json recipe.json
 
-# --allow-multiple-definition: brotli 3.x (pingora, ffi-api default) and
-# brotli 6.x (deno_web, ffi-api explicit) both export identical C FFI symbols.
-# GNU ld rejects duplicates; macOS ld64 allows them silently. The symbols are
-# functionally identical so picking either definition is safe.
-#
 # Cook dependencies only -- this layer is cached until Cargo.toml/Cargo.lock change.
-RUN RUSTFLAGS="-C link-args=-Wl,--allow-multiple-definition" \
-    cargo chef cook --release --recipe-path recipe.json -p gateway-core
+RUN cargo chef cook --release --recipe-path recipe.json -p gateway-core
 
 COPY Cargo.toml Cargo.lock ./
 COPY gateway-core/ gateway-core/
 COPY openapi-overlay/ openapi-overlay/
 COPY opengateway-js-runtime/ opengateway-js-runtime/
-COPY --from=js-builder /usr/src/opengateway/gateway-core/js/dist/ gateway-core/js/dist/
+COPY opengateway-sandbox/ opengateway-sandbox/
 
-RUN RUSTFLAGS="-C link-args=-Wl,--allow-multiple-definition" cargo build --release -p gateway-core
+RUN cargo build --release -p gateway-core
 
 FROM node:22-bookworm-slim
