--- interceptors/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "opengateway-interceptors",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "build": "node esbuild.mjs"
-  },
-  "devDependencies": {
-    "esbuild": "^0.25.0",
-    "@types/node": "^22.0.0",
-    "typescript": "^5.0.0"
-  },
-  "dependencies": {
-    "ajv": "^8.0.0"
-  }
-}
