+++ opengateway-js-runtime/node-runtime/package.json
@@ -4,6 +4,7 @@
   "description": "Out-of-process Node.js plugin server for OpenGateway",
   "private": true,
   "dependencies": {
+    "ajv": "^8.17.1",
     "msgpackr": "^1.11.0"
   }
 }
